power: qpnp-fg-gen4: Add support for CYCLE_COUNT property
Currently, CYCLE_COUNTS property shows the bucket cycle
counters as a string. Add CYCLE_COUNT property in FG that can
provide an average of all the bucket counters.

Move the helper functions that gets the cycle_counts and
cycle_count property to fg-alg.c so that they can be re-used.

Change-Id: I17291cb6a3c15fe66700f6ee1bbb3dd7327ce6f2
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>