power: qpnp-fg-gen4: Update whitelisting values for profile integrity bit
Profile integrity word holds the data that indicates whether the
bootloader or HLOS has loaded the profile or an SOC estimate is
done etc. So far bits 0:3 are used for indicating this. Bit 4 is
added to represent whether the profile is loaded for the first
time. This can be used later to clear SOH data stored by the
userspace.

Change-Id: I9bb39d2df3d948fe4253b4241a02db28053d8346
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>